make request and response
