+++ Cargo.toml
@@ -12,6 +12,7 @@ rust-ini = "0.18.0"
 jsonrpc-ipc-server = "18.0.0"
 jsonrpc = {version = "0.14.0", features = ["simple_uds"]}
 serde_json = "1.0.91"
+serde = { version = "1.0.152", features = ["derive"] }
 
 [patch.crates-io]
 rust-ini = { git = "https://github.com/exciting-supervisord/rust-ini.git" }
