+++ src/lib/lib.rs
@@ -1,4 +1,4 @@
 pub mod config;
-pub mod output_message;
 pub mod process_status;
-pub mod rpc_error;
+pub mod request;
+pub mod response;
