--- src/lib/output_message.rs
@@ -1,10 +1,0 @@
-pub struct OutputMessage {
-    name: String,
-    message: String,
-}
-
-impl std::fmt::Display for OutputMessage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}: {}", self.name, self.message)
-    }
-}
