+++ src/lib/request.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Request {
+    method: String,
+    args: Vec<String>,
+}
